crud product update

+++ src/main/java/com/ute/shop/controller/admin/ProductController.java
@@ -1,8 +1,9 @@
 package com.ute.shop.controller.admin;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
-import java.util.function.Function;
+import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -10,10 +11,13 @@ import javax.validation.Valid;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.Resource;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
@@ -25,6 +29,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.ute.shop.domain.Category;
@@ -33,6 +38,7 @@ import com.ute.shop.model.CategoryDto;
 import com.ute.shop.model.ProductDto;
 import com.ute.shop.service.CategoryService;
 import com.ute.shop.service.ProductService;
+import com.ute.shop.service.StorageService;
 
 
 @Controller
@@ -43,6 +49,9 @@ public class ProductController {
 	CategoryService categoryService;
 	@Autowired
 	ProductService productService;
+	@Autowired
+	StorageService storageService;
+	
 	@ModelAttribute("categories")
 	public List<CategoryDto> getCategories(){
 		return categoryService.findAll().stream().map(item->{
@@ -54,7 +63,9 @@ public class ProductController {
 	
 	@GetMapping("add")
 	public String add(Model model) {
-		model.addAttribute("product", new ProductDto());
+		ProductDto dto = new ProductDto();
+		dto.setIsEdit(true);
+		model.addAttribute("product", dto);
 		return "admin/products/addOrEdit";
 	}
 	@GetMapping("edit/{productId}")
@@ -64,29 +75,57 @@ public class ProductController {
 		if(optional.isPresent()) {
 			Product entity = optional.get();
 			BeanUtils.copyProperties(entity, productDto);
-//			productDto.setIsEdit(true);
+			productDto.setCategoryId(entity.getCategory().getCategoryId());
+			productDto.setIsEdit(true);
 			model.addAttribute("product",productDto);
 			return new ModelAndView("admin/products/addOrEdit",model);
 		}
 		model.addAttribute("message","Product is not exited");
 		return new ModelAndView( "forward:/admin/products/",model);
 	}
+	@GetMapping("/images/{fileName:.+}")
+	@ResponseBody
+	public ResponseEntity<Resource> serveFile(@PathVariable String fileName){
+		Resource file = storageService.loadAsResource(fileName);
+		return ResponseEntity.ok().header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + file.getFilename() + "\"").body(file);
+	}
+	
 	@GetMapping("delete/{productId}")
-	public ModelAndView delete(ModelMap model, @PathVariable("productId") Integer productId) {
+	public ModelAndView delete(ModelMap model, @PathVariable("productId") Integer productId) throws IOException {
+		Optional<Product> optional = productService.findById(productId);
+		if(optional.isPresent()) {
+			if(!StringUtils.isEmpty(optional.get().getImage()) ) {
+				storageService.delete(optional.get().getImage());
+			}
+			productService.deleteById(productId);
+			model.addAttribute("message", "Product is Deleted !");
+		}
+		else {
+			model.addAttribute("message", "Product is not Found !");
+		}
 		
-		productService.deleteById(productId);
-		model.addAttribute("message", "Product is Deleted !");
 		return new ModelAndView("forward:/admin/products",model);
 	}
 	
 	@PostMapping("saveOrUpdate")
 	public ModelAndView saveOrUpdate(ModelMap model,@Valid @ModelAttribute("product") ProductDto productDto, BindingResult bindingResult) {
 		if(bindingResult.hasErrors()) {
-			bindingResult.getAllErrors().forEach(item -> System.out.println(item));
 			return new ModelAndView("admin/products/addOrEdit");
 		}
 		Product entity = new Product();
 		BeanUtils.copyProperties(productDto, entity);
+		
+		Category category = new Category();
+		category.setCategoryId(productDto.getCategoryId());
+		entity.setCategory(category);
+		
+		if(!productDto.getImageFile().isEmpty()) {
+			UUID uuid = UUID.randomUUID();
+			String uuString = uuid.toString();
+			entity.setImage(storageService.getStoredFileName(productDto.getImageFile(), uuString));
+			storageService.store(productDto.getImageFile(), entity.getImage());
+		}
+		
 		productService.save(entity);
 		model.addAttribute("message", "Product is Saved !");
 		return new ModelAndView("forward:/admin/products", model);
